% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiR6functions.R
\name{extractTS}
\alias{extractTS}
\title{Extract fMRI time-series}
\usage{
extractTS(niftiObject, maskData = NULL, xyzloc = NULL)
}
\arguments{
\item{maskData}{object of class \code{\link{niftiDataR6}} containing 3D mask}

\item{xyzloc}{matrix with x,y,z locations of to-be-extracted data}

\item{niftiData}{object of class\code{\link{niftiDataR6}} containing 4D time-series}
}
\value{
matrix of voxel by time
}
\description{
Extract time-series and compile in a voxel by time matrix
}
