% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiR6class.R
\docType{class}
\name{niftiDataR6}
\alias{niftiDataR6}
\title{Class with file information of NIfTI files}
\format{\code{\link{R6Class}} object.}
\usage{
niftiDataR6
}
\value{
Object of \code{\link{R6Class}} with info, header and data of a NIfTI file
}
\description{
Class with file information of NIfTI files
}
\section{Fields}{

\describe{
\item{\code{data}}{array of data}
}}

\section{Methods}{

\describe{
  \item{\code{show()}}{Shows the niftiData object.}
  \item{\code{readData()}}{Reads in the data (+header)from a NIfTI file.}
  \item{\code{writeData()}}{Writes data (+header) to a NIfTI file.}
  \item{\code{setData()}}{Adds an entire array/matrix/vector to the data slot with checks.}
}
}

\examples{
niftiData <- niftiDataR6$new()

}
\keyword{data}
