% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiR6class.R
\docType{class}
\name{niftiHeaderR6}
\alias{niftiHeaderR6}
\title{Class with header information from a NIfTI file}
\format{\code{\link{R6Class}} object.}
\usage{
niftiHeaderR6
}
\value{
Object of \code{\link{R6Class}} with information on a NIfTI file
}
\description{
Class with header information from a NIfTI file
}
\section{Fields}{

\describe{
\item{\code{sizeof_hdr}}{nifti header slot}

\item{\code{data_type}}{nifti header slot}

\item{\code{db_name}}{nifti header slot}

\item{\code{extents}}{nifti header slot}

\item{\code{session_error}}{nifti header slot}

\item{\code{regular}}{nifti header slot}

\item{\code{dim_info}}{nifti header slot}

\item{\code{dims}}{nifti header slot}

\item{\code{intent_p1}}{nifti header slot}

\item{\code{intent_p2}}{nifti header slot}

\item{\code{intent_p3}}{nifti header slot}

\item{\code{intent_code}}{nifti header slot}

\item{\code{datatype}}{nifti header slot}

\item{\code{bitpix}}{nifti header slot}

\item{\code{slice_start}}{nifti header slot}

\item{\code{pixdim}}{nifti header slot}

\item{\code{vox_offset}}{nifti header slot}

\item{\code{scl_slope}}{nifti header slot}

\item{\code{scl_inter}}{nifti header slot}

\item{\code{slice_end}}{nifti header slot}

\item{\code{slice_code}}{nifti header slot}

\item{\code{xyzt_units}}{nifti header slot}

\item{\code{cal_max}}{nifti header slot}

\item{\code{cal_min}}{nifti header slot}

\item{\code{slice_duration}}{nifti header slot}

\item{\code{toffset}}{nifti header slot}

\item{\code{glmax}}{nifti header slot}

\item{\code{glmin}}{nifti header slot}

\item{\code{descrip}}{nifti header slot}

\item{\code{aux_file}}{nifti header slot}

\item{\code{qform_code}}{nifti header slot}

\item{\code{sform_code}}{nifti header slot}

\item{\code{quatern_b}}{nifti header slot}

\item{\code{quatern_c}}{nifti header slot}

\item{\code{quatern_d}}{nifti header slot}

\item{\code{qoffset_x}}{nifti header slot}

\item{\code{qoffset_y}}{nifti header slot}

\item{\code{qoffset_z}}{nifti header slot}

\item{\code{srow_x}}{nifti header slot}

\item{\code{srow_y}}{nifti header slot}

\item{\code{srow_z}}{nifti header slot}

\item{\code{intent_name}}{nifti header slot}

\item{\code{magic}}{nifti header slot}

\item{\code{extension_char}}{nifti header slot}

\item{\code{extension_data}}{nifti header slot}
}}

\section{Methods}{

\describe{
  \item{\code{show()}}{Shows header info}
  \item{\code{readHeader(filename)}}{Reads in header info}
  \item{\code{writeHeader()}}{Writes header info to file}
  \item{\code{changeDatatype(datatype)}}{sets the data type + bitpix according a nifti-code (numeric or string)}


}
}

\examples{
niftiHeader <- niftiHeaderR6$new()

}
\keyword{data}
