% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiR6class.R
\docType{class}
\name{niftiInfoR6}
\alias{niftiInfoR6}
\title{Class with file information of NIfTI files}
\format{\code{\link{R6Class}} object.}
\usage{
niftiInfoR6
}
\value{
Object of \code{\link{R6Class}} with information on a NIfTI file
}
\description{
Class with file information of NIfTI files
}
\section{Fields}{

\describe{
\item{\code{fullpath}}{Full path of the file}

\item{\code{filename}}{Name of the file (without extension)}

\item{\code{extension}}{Extension of the file}

\item{\code{gzipped}}{Is the file compressed?}

\item{\code{endian}}{Endianness of the file}

\item{\code{data.type}}{Type of data in file}

\item{\code{data.signed}}{Are data.type signed?}

\item{\code{connection}}{Connection to the file}
}}

\section{Methods}{

\describe{
  \item{\code{getFullPath()}}{Returns string of path to the file.}
  \item{\code{changeName()}}{Changes the name of the file in the R6 object.}
  \item{\code{changePath()}}{Changes the path of the file in the R6 object.}
  \item{\code{getName()}}{Returns string of the filename.}
  \item{\code{showInfo()}}{Shows the info of the file.}

}
}

\examples{
niftiInfo <- niftiInfoR6$new()
}
\keyword{data}
