% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiR6functions.R
\name{plotNifti}
\alias{plotNifti}
\title{plot a NIfTI-1/NIfTI-2 object}
\usage{
plotNifti(niftiObject, type = c("sag", "axi", "cor"), volume = "avg",
  color = c("grayscale", "red-yellow"), asp = 1)
}
\arguments{
\item{niftiObject}{niftiR6 object}
}
\description{
plots a NIfTI-1/NIfTI-2 object
}
