% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiR6functions.R
\name{readNifti}
\alias{readNifti}
\title{read in a NIfTI-1/NIfTI-2 file}
\usage{
readNifti(filename = NULL)
}
\arguments{
\item{filename}{name of the file to open (may include path)}
}
\value{
object of class \code{\link{niftiDataR6}}
}
\description{
Reads in a file and returns an object of niftiDataR6 class
}
