% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiR6functions.R
\name{writeNifti}
\alias{writeNifti}
\title{write to a NIfTI-1/NIfTI-2 file}
\usage{
writeNifti(niftiObject, overwriteWarning = TRUE)
}
\arguments{
\item{niftiObject}{object of class \code{\link{niftiDataR6}}}

\item{overwriteWarning}{warn (with a user option to override) when about to overwrite a file}
}
\value{
object of class \code{\link{niftiDataR6}}
}
\description{
writes niftiDataR6 object to the file specified in the niftiDataR6object
or to a file provided by filename.
}
